refactor: ReviewDocument의 하위 객체들 별도 생성
